Add files via upload
This file contains a shortened version of the R code for a class project in ECMA3132: Applications of Econometric and Data Science Methods at the University of Chicago. In it I extracted relevant information from a large number of data files from the BLS, cleaned the raw data, and performed DID regression on it.